+++ src/taglib.pyx
@@ -123,7 +123,7 @@ cdef class File:
         ValueError
             When attempting to save after the file was closed.
         """
-        if not self.cFile:
+        if self.is_closed:
             raise ValueError('I/O operation on closed file.')
         if self.readOnly:
             raise OSError(f'Unable to save tags: file is read-only')
@@ -159,42 +159,47 @@ cdef class File:
         """Closes the file by deleting the underlying Taglib::File object. This will close any open
         streams. Calling methods like `save()` or the read-only properties after `close()` will
         raise an exception."""
+        if self.is_closed:
+            raise ValueError("File already closed")
         del self.cFile
         self.cFile = NULL
 
     def __dealloc__(self):
         if self.cFile:
             del self.cFile
-        
+
+    @property
+    def is_closed(self):
+        return self.cFile is NULL
+
     property length:
         def __get__(self):
-            if not self.cFile:
-                raise ValueError('I/O operation on closed file.')
+            self.check_closed()
             return self.cFile.audioProperties().length()
             
     property bitrate:
         def __get__(self):
-            if not self.cFile:
-                raise ValueError('I/O operation on closed file.')
+            self.check_closed()
             return self.cFile.audioProperties().bitrate()
     
     property sampleRate:
         def __get__(self):
-            if not self.cFile:
-                raise ValueError('I/O operation on closed file.')
+            self.check_closed()
             return self.cFile.audioProperties().sampleRate()
             
     property channels:
         def __get__(self):
-            if not self.cFile:
-                raise ValueError('I/O operation on closed file.')
+            self.check_closed()
             return self.cFile.audioProperties().channels()
     
     property readOnly:
         def __get__(self):
-            if not self.cFile:
-                raise ValueError('I/O operation on closed file.')
+            self.check_closed()
             return self.cFile.readOnly()
-        
+
+    cdef check_closed(self):
+        if self.is_closed:
+            raise ValueError('I/O operation on closed file.')
+
     def __repr__(self):
         return f"File('{self.path}')"